Implement loading DDS textures at run-time
Based on 34ab1c8a36e61b386c210fb908b9ebfa86513149
