+++ modules/dds/image_loader_dds.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  register_types.cpp                                                    */
+/*  image_loader_dds.h                                                    */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             REDOT ENGINE                               */
@@ -30,34 +30,65 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#include "register_types.h"
+#ifndef IMAGE_LOADER_DDS_H
+#define IMAGE_LOADER_DDS_H
 
-#include "image_loader_dds.h"
-#include "texture_loader_dds.h"
+#include "core/io/image_loader.h"
 
-static Ref<ResourceFormatDDS> resource_loader_dds;
-static Ref<ImageLoaderDDS> image_loader_dds;
+class ImageLoaderDDS : public ImageFormatLoader {
+public:
+	// The legacy bitmasked format names here represent the actual data layout in the files,
+	// while their official names are flipped (e.g. RGBA8 layout is officially called ABGR8).
+	enum DDSFormat {
+		DDS_DXT1,
+		DDS_DXT3,
+		DDS_DXT5,
+		DDS_ATI1,
+		DDS_ATI2,
+		DDS_BC6U,
+		DDS_BC6S,
+		DDS_BC7,
+		DDS_R16F,
+		DDS_RG16F,
+		DDS_RGBA16F,
+		DDS_R32F,
+		DDS_RG32F,
+		DDS_RGB32F,
+		DDS_RGBA32F,
+		DDS_RGB9E5,
+		DDS_RGB8,
+		DDS_RGBA8,
+		DDS_BGR8,
+		DDS_BGRA8,
+		DDS_BGR5A1,
+		DDS_BGR565,
+		DDS_B2GR3,
+		DDS_B2GR3A8,
+		DDS_BGR10A2,
+		DDS_RGB10A2,
+		DDS_BGRA4,
+		DDS_LUMINANCE,
+		DDS_LUMINANCE_ALPHA,
+		DDS_LUMINANCE_ALPHA_4,
+		DDS_MAX
+	};
 
-void initialize_dds_module(ModuleInitializationLevel p_level) {
-	if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
-		return;
-	}
+	enum DDSType {
+		DDST_2D = 1,
+		DDST_CUBEMAP,
+		DDST_3D,
 
-	resource_loader_dds.instantiate();
-	ResourceLoader::add_resource_format_loader(resource_loader_dds);
+		DDST_TYPE_MASK = 0x7F,
+		DDST_ARRAY = 0x80,
+	};
 
-	image_loader_dds.instantiate();
-	ImageLoader::add_image_format_loader(image_loader_dds);
-}
+	static Error load_layer(Ref<Image> p_layer, Ref<FileAccess> p_file, DDSFormat p_dds_format, uint32_t p_width, uint32_t p_height, uint32_t p_mipmaps, uint32_t p_pitch, uint32_t p_flags, Vector<uint8_t> &r_src_data);
+	static Error load_image_layers(Vector<Ref<Image>> &p_layers, Ref<FileAccess> p_file, uint32_t *r_dds_type = nullptr);
 
-void uninitialize_dds_module(ModuleInitializationLevel p_level) {
-	if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
-		return;
-	}
+	virtual Error load_image(Ref<Image> p_image, Ref<FileAccess> f, BitField<ImageFormatLoader::LoaderFlags> p_flags, float p_scale);
+	virtual void get_recognized_extensions(List<String> *p_extensions) const;
+	virtual bool should_import(const String &p_resource_type) const;
+	ImageLoaderDDS();
+};
 
-	ResourceLoader::remove_resource_format_loader(resource_loader_dds);
-	resource_loader_dds.unref();
-
-	ImageLoader::remove_image_format_loader(image_loader_dds);
-	image_loader_dds.unref();
-}
+#endif // IMAGE_LOADER_DDS_H
