+++ tests/core/io/test_image.h
@@ -125,6 +125,19 @@ TEST_CASE("[Image] Saving and loading") {
 			"The BMP image should load successfully.");
 #endif // MODULE_BMP_ENABLED
 
+#ifdef MODULE_DDS_ENABLED
+	// Load DDS
+	Ref<Image> image_dds = memnew(Image());
+	Ref<FileAccess> f_dds = FileAccess::open(TestUtils::get_data_path("images/icon.dds"), FileAccess::READ, &err);
+	REQUIRE(f_dds.is_valid());
+	PackedByteArray data_dds;
+	data_dds.resize(f_dds->get_length() + 1);
+	f_dds->get_buffer(data_dds.ptrw(), f_dds->get_length());
+	CHECK_MESSAGE(
+			image_dds->load_dds_from_buffer(data_dds) == OK,
+			"The DDS image should load successfully.");
+#endif // MODULE_DDS_ENABLED
+
 #ifdef MODULE_JPG_ENABLED
 	// Load JPG
 	Ref<Image> image_jpg = memnew(Image());
